refactor: update files in mobile app module using new package definition for user preferences repository
